Remove some Python versions from Travis testing
This should get CI jobs passing again. pypy3 is definitely a Travis
failure. It doesn't seem to be working for anybody right now.

pypy and python2.6 have an issue with SSL. The error message
indicates it's due to the old version of Python being used but I
haven't dug deeper yet to find out if this is fixable.